// migrations/snapshots/46c7ebca3b5817bbd8dc6251871f457f09f0d18f14e085195f5a84d6ebd09e1b/contract.d.ts
export type CodecTypes = PgTypes;
export type LaneCodecTypes = CodecTypes;
export type QueryOperationTypes = PgAdapterQueryOps<CodecTypes>;
// ...
export type FieldOutputTypes = {
  readonly public: {
    readonly Budget: {
      readonly id: CodecTypes['pg/text@1']['output'];
      readonly limitAmount: CodecTypes['pg/numeric@1']['output'];
      readonly month: CodecTypes['pg/int4@1']['output'];
      readonly year: CodecTypes['pg/int4@1']['output'];
      readonly categoryId: CodecTypes['pg/text@1']['output'];
      readonly userId: CodecTypes['pg/text@1']['output'];
    };
    readonly Category: {
      readonly id: CodecTypes['pg/text@1']['output'];
      readonly name: CodecTypes['pg/text@1']['output'];
      readonly type: 'EXPENSE' | 'INCOME' | 'INVEST' | 'DEBT' | 'SAVINGS';
      readonly color: CodecTypes['pg/text@1']['output'];
      readonly icon: CodecTypes['pg/text@1']['output'] | null;
      readonly isDefault: CodecTypes['pg/bool@1']['output'];
      readonly createdAt: CodecTypes['pg/timestamptz-temporal@1']['output'];
      readonly userId: CodecTypes['pg/text@1']['output'];
    };
    readonly Debt: {
      readonly id: CodecTypes['pg/text@1']['output'];
      readonly person: CodecTypes['pg/text@1']['output'];
      readonly amount: CodecTypes['pg/numeric@1']['output'];
      readonly paidAmount: CodecTypes['pg/numeric@1']['output'];
      readonly direction: 'OWE' | 'OWED';
      readonly status: 'PENDING' | 'PARTIAL' | 'PAID';
      readonly dueDate: CodecTypes['pg/timestamptz-temporal@1']['output'] | null;
      readonly note: CodecTypes['pg/text@1']['output'] | null;
      readonly createdAt: CodecTypes['pg/timestamptz-temporal@1']['output'];
      readonly updatedAt: CodecTypes['pg/timestamptz-temporal@1']['output'];
      readonly userId: CodecTypes['pg/text@1']['output'];
    };
    readonly Goal: {
      readonly id: CodecTypes['pg/text@1']['output'];
      readonly name: CodecTypes['pg/text@1']['output'];
      readonly targetAmount: CodecTypes['pg/numeric@1']['output'];
      readonly savedAmount: CodecTypes['pg/numeric@1']['output'];
      readonly deadline: CodecTypes['pg/timestamptz-temporal@1']['output'] | null;
      readonly note: CodecTypes['pg/text@1']['output'] | null;
      readonly createdAt: CodecTypes['pg/timestamptz-temporal@1']['output'];
      readonly updatedAt: CodecTypes['pg/timestamptz-temporal@1']['output'];
      readonly userId: CodecTypes['pg/text@1']['output'];
    };
    readonly InvestLog: {
      readonly id: CodecTypes['pg/text@1']['output'];
      readonly action: 'BUY' | 'SELL';
      readonly quantity: CodecTypes['pg/numeric@1']['output'];
      readonly price: CodecTypes['pg/numeric@1']['output'];
      readonly recordedAt: CodecTypes['pg/timestamptz-temporal@1']['output'];
      readonly createdAt: CodecTypes['pg/timestamptz-temporal@1']['output'];
      readonly investmentId: CodecTypes['pg/text@1']['output'];
    };
    readonly Investment: {
      readonly id: CodecTypes['pg/text@1']['output'];
      readonly name: CodecTypes['pg/text@1']['output'];
      readonly ticker: CodecTypes['pg/text@1']['output'] | null;
      readonly quantity: CodecTypes['pg/numeric@1']['output'];
      readonly buyPrice: CodecTypes['pg/numeric@1']['output'];
      readonly currentPrice: CodecTypes['pg/numeric@1']['output'] | null;
      readonly boughtAt: CodecTypes['pg/timestamptz-temporal@1']['output'];
      readonly createdAt: CodecTypes['pg/timestamptz-temporal@1']['output'];
      readonly updatedAt: CodecTypes['pg/timestamptz-temporal@1']['output'];
      readonly userId: CodecTypes['pg/text@1']['output'];
    };
    readonly Transaction: {
      readonly id: CodecTypes['pg/text@1']['output'];
      readonly amount: CodecTypes['pg/numeric@1']['output'];
      readonly note: CodecTypes['pg/text@1']['output'] | null;
      readonly recordedAt: CodecTypes['pg/timestamptz-temporal@1']['output'];
      readonly createdAt: CodecTypes['pg/timestamptz-temporal@1']['output'];
      readonly updatedAt: CodecTypes['pg/timestamptz-temporal@1']['output'];
      readonly type: 'INCOME' | 'EXPENSE';
      readonly categoryId: CodecTypes['pg/text@1']['output'];
      readonly goalId: CodecTypes['pg/text@1']['output'] | null;
      readonly userId: CodecTypes['pg/text@1']['output'];
    };
    readonly User: {
      readonly id: CodecTypes['pg/text@1']['output'];
      readonly email: CodecTypes['pg/text@1']['output'];
      readonly name: CodecTypes['pg/text@1']['output'] | null;
      readonly password: CodecTypes['pg/text@1']['output'];
      readonly timezone: CodecTypes['pg/text@1']['output'];
      readonly createdAt: CodecTypes['pg/timestamptz-temporal@1']['output'];
      readonly updatedAt: CodecTypes['pg/timestamptz-temporal@1']['output'];
    };
  };
};
export type FieldInputTypes = {
  readonly public: {
    readonly Budget: {
      readonly id: CodecTypes['pg/text@1']['input'];
      readonly limitAmount: CodecTypes['pg/numeric@1']['input'];
      readonly month: CodecTypes['pg/int4@1']['input'];
      readonly year: CodecTypes['pg/int4@1']['input'];
      readonly categoryId: CodecTypes['pg/text@1']['input'];
      readonly userId: CodecTypes['pg/text@1']['input'];
    };
    readonly Category: {
      readonly id: CodecTypes['pg/text@1']['input'];
      readonly name: CodecTypes['pg/text@1']['input'];
      readonly type: 'EXPENSE' | 'INCOME' | 'INVEST' | 'DEBT' | 'SAVINGS';
      readonly color: CodecTypes['pg/text@1']['input'];
      readonly icon: CodecTypes['pg/text@1']['input'] | null;
      readonly isDefault: CodecTypes['pg/bool@1']['input'];
      readonly createdAt: CodecTypes['pg/timestamptz-temporal@1']['input'];
      readonly userId: CodecTypes['pg/text@1']['input'];
    };
    readonly Debt: {
      readonly id: CodecTypes['pg/text@1']['input'];
      readonly person: CodecTypes['pg/text@1']['input'];
      readonly amount: CodecTypes['pg/numeric@1']['input'];
      readonly paidAmount: CodecTypes['pg/numeric@1']['input'];
      readonly direction: 'OWE' | 'OWED';
      readonly status: 'PENDING' | 'PARTIAL' | 'PAID';
      readonly dueDate: CodecTypes['pg/timestamptz-temporal@1']['input'] | null;
      readonly note: CodecTypes['pg/text@1']['input'] | null;
      readonly createdAt: CodecTypes['pg/timestamptz-temporal@1']['input'];
      readonly updatedAt: CodecTypes['pg/timestamptz-temporal@1']['input'];
      readonly userId: CodecTypes['pg/text@1']['input'];
    };
    readonly Goal: {
      readonly id: CodecTypes['pg/text@1']['input'];
      readonly name: CodecTypes['pg/text@1']['input'];
      readonly targetAmount: CodecTypes['pg/numeric@1']['input'];
      readonly savedAmount: CodecTypes['pg/numeric@1']['input'];
      readonly deadline: CodecTypes['pg/timestamptz-temporal@1']['input'] | null;
      readonly note: CodecTypes['pg/text@1']['input'] | null;
      readonly createdAt: CodecTypes['pg/timestamptz-temporal@1']['input'];
      readonly updatedAt: CodecTypes['pg/timestamptz-temporal@1']['input'];
      readonly userId: CodecTypes['pg/text@1']['input'];
    };
    readonly InvestLog: {
      readonly id: CodecTypes['pg/text@1']['input'];
      readonly action: 'BUY' | 'SELL';
      readonly quantity: CodecTypes['pg/numeric@1']['input'];
      readonly price: CodecTypes['pg/numeric@1']['input'];
      readonly recordedAt: CodecTypes['pg/timestamptz-temporal@1']['input'];
      readonly createdAt: CodecTypes['pg/timestamptz-temporal@1']['input'];
      readonly investmentId: CodecTypes['pg/text@1']['input'];
    };
    readonly Investment: {
      readonly id: CodecTypes['pg/text@1']['input'];
      readonly name: CodecTypes['pg/text@1']['input'];
      readonly ticker: CodecTypes['pg/text@1']['input'] | null;
      readonly quantity: CodecTypes['pg/numeric@1']['input'];
      readonly buyPrice: CodecTypes['pg/numeric@1']['input'];
      readonly currentPrice: CodecTypes['pg/numeric@1']['input'] | null;
      readonly boughtAt: CodecTypes['pg/timestamptz-temporal@1']['input'];
      readonly createdAt: CodecTypes['pg/timestamptz-temporal@1']['input'];
      readonly updatedAt: CodecTypes['pg/timestamptz-temporal@1']['input'];
      readonly userId: CodecTypes['pg/text@1']['input'];
    };
    readonly Transaction: {
      readonly id: CodecTypes['pg/text@1']['input'];
      readonly amount: CodecTypes['pg/numeric@1']['input'];
      readonly note: CodecTypes['pg/text@1']['input'] | null;
      readonly recordedAt: CodecTypes['pg/timestamptz-temporal@1']['input'];
      readonly createdAt: CodecTypes['pg/timestamptz-temporal@1']['input'];
      readonly updatedAt: CodecTypes['pg/timestamptz-temporal@1']['input'];
      readonly type: 'INCOME' | 'EXPENSE';
      readonly categoryId: CodecTypes['pg/text@1']['input'];
      readonly goalId: CodecTypes['pg/text@1']['input'] | null;
      readonly userId: CodecTypes['pg/text@1']['input'];
    };
    readonly User: {
      readonly id: CodecTypes['pg/text@1']['input'];
      readonly email: CodecTypes['pg/text@1']['input'];
      readonly name: CodecTypes['pg/text@1']['input'] | null;
      readonly password: CodecTypes['pg/text@1']['input'];
      readonly timezone: CodecTypes['pg/text@1']['input'];
      readonly createdAt: CodecTypes['pg/timestamptz-temporal@1']['input'];
      readonly updatedAt: CodecTypes['pg/timestamptz-temporal@1']['input'];
    };
  };
};
// ...
export type TypeMaps = TypeMapsType<
  CodecTypes,
  QueryOperationTypes,
  FieldOutputTypes,
  FieldInputTypes,
  StorageColumnTypes,
  StorageColumnInputTypes,
  AggregateTypes
>;
// ...
export type Contract = ContractWithTypeMaps<ContractBase, TypeMaps>;